Move integrand_slq_spd from slq.py to lanczos.py and delete empty slq.py and test_slq.py content
